requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.3.26
python-dotenv==1
# Frontend
streamlit==1.28.1